feat(api): deliver events written by another process on the same database
- Tail the append-only events log and publish each row through the
  existing notifier, so a change written by a second product reaches
  this process's subscribers without a broker or shared code
- Re-read below the high-water mark for a grace period: AUTO_INCREMENT
  assigns an id before the row is visible, so a cursor that jumps to the
  highest id seen steps over a transaction committing after it
- Suppress the repeats that re-reading produces, and skip rows this
  process appended itself so a local write does not invalidate twice
- Widen the notify hook's event id to uint64 to match events.id, which
  is BIGINT UNSIGNED; the previous cast would silently rewrite
  notifications.source_event_id, the key behind at-least-once delivery
- Start at the end of the log so a restart does not replay history
- Record the id-visibility hazard and the at-least-once guarantee in the
  core protocol

+++ apps/flow-api/internal/ai/agentruntime/cron.go
@@ -27,7 +27,7 @@ type Job struct {
 	// uses it to look up the source event's task_id so the emitted
 	// ai.agent.run.* events stay bound to the same task and surface in
 	// the task's run-history view.
-	SourceEventID uint32
+	SourceEventID uint64
 	// DedupeKey is the agent_runs.dedupe_key value that the scheduler /
 	// enqueuer assigned to this job. The runner uses it to recover the
 	// dispatch hint (signal id for the signaljudge path, currently)
