+++ apps/flow-api/internal/mcp/sse.go
@@ -230,7 +230,7 @@ func buildEventNotification(eventType string, seq int64) string {
 // events to SSE-connected MCP clients. It must be non-blocking. The
 // eventInternalID parameter is accepted for signature compatibility
 // but unused — MCP clients receive only the type and seq.
-func (h *Handler) onWorkspaceEvent(ctx context.Context, workspaceID uint32, eventType string, _ uint32) {
+func (h *Handler) onWorkspaceEvent(ctx context.Context, workspaceID uint32, eventType string, _ uint64) {
 	seq := eventbus.SeqFromContext(ctx)
 	h.sse.broadcast(workspaceID, sseEvent{
 		EventType: "workspace.event",
